ImportRelations - re-add csv support EEE XXX _is_merger
Took the logic from _write_record - so as if the added and changed tags
were applied. I had to hack _read_record_sp - why are we doing:

relations = self.id_stored_data[b'FACT'][record.fid]

there?

This should work but do we want to add it? We should also doc it EEE
This has melted my brain - I really can't follow what the desired
result is in read_record_sp. We probably want to override what's
in id_stored_data for the record.fid - _Mod_Export_Link is the only
case where we might read multiple mods so in exporting FactionRelations
the relation arrays would be merged while for ActorFactions the last
loaded mod that had record.fid would populate those factions array
attributes it had. Now in the patcher FactionRelations were broken,
while _write_record (the other use of read_record_sp) is not used
in the patcher. We use it now for FactionRelations, effectively adding
and changing relations to the record at hand based on the csv, but
not removing relations that are not present.

If FactionRelations should merge let's make this explicit - still not
clear what _write_record should check - note that now it checks for
differences but only changes/additions are written