Fixup for converters fullRefresh:
We clean bcf_scd so we would get a KeyError in path_crc = bcf_scd[bcfPath]
which would also be swallowed in the bare except - made this more specific.
StateError should not be used in place of FileNotFoundError.

Mopy/bash/basher/installer_links.py: cleanup and use _SingleInstallable,
takes down a few is_marker. Note that _RefreshingLink is-a
_SingleInstallable so no need for the _enable override.

_refresh_from_data_dir is only called once and both params are passed

Mopy/bash/basher/installers_links.py: prune some data_sizeCrcDate and
simplify a bit (more explicit to operate with keys rather than the set
of keys and should be equally fast).

belt: self.bArchive can be used here - even better we need an Installer
method - but first I need to figure out what we are trying to do and
why we use this walk(relative=True) (yak)