Remove last balt import from BAIN:
Ugly means to a laudable end - if we need more of this we should come up
with a more elegant solution but this will do for now (ideally for ever)

Come up with a better naming for gui/balt callbacks - or not we don't
want any more of this. Note some of it is around dlls - SEPlugins tab
could to the rescue? And an open question:

@@ -672,3 +672,3 @@ def __skipExecutable(checkOBSE, fileLower, full, archiveRoot, dll_size,
                                         desc, ext, badDlls, goodDlls)
-            if not askYes(balt.Link.Frame, message, dialogTitle):
+            if not askYes(None, message, dialogTitle): ##: was balt.Link.Frame <=> None?
                 badDlls[fileLower].append([archiveRoot, dll_size, crc])