Fix balt.sizes not getting saved
This must be a regression, there's no way this has been defunct since
forever. No idea when it broke though. The problem is that we were
creating a new dict for balt.sizes each time. Could have handled this
by implementing a setdefault for DataDict/Settings, but I opted to
instead add a default and move the balt.sizes init after the
loadDefault. I don't *think* this can cause issues.

Utumno: see discussion in
https://github.com/wrye-bash/wrye-bash/commit/a1a755fadadfce033ddc0a926ab170aac4638f8f
No regression actually, this was never saved - but as discussed above
let's save it now.

Poc move sizes into settings:

To really finalize this we would need to split _Log into two classes
inheriting from WindowFrame/Dialog and a common mixin for the common code
(maybe). This needs hunting down the many uses and replace the asDialog
with the classes - but that's more than poc (edit: done later in this
branch)

For now I got us rid of _base_key and the sizes dict - note the
duplicate

        key__pos_ = f'{self._settings_key}.pos'
        key__size_ = f'{self._settings_key}.size'

in WindowFrame and _Log (duh)

Co-authored-by: MrD <the.ubik@gmail.com>