Remove the mouse capture code - drag and drop works fine on __WXMAC__: TTT
We need to test this on windows but miraculously works on mac and maybe
also in __WXGTK__?
What was happening is that somehow the event.Skip() in OnDragStart was
propagating the event which somehow blocked the OnDrag from being called

Removing the event.Skip() made drag and drop work but the button would
not be pushed for the life of me. Then I realized that the OnDragEnd
could be hijacked to call the button.Execute anyways. Maybe we could
return FINISH from on_clicked listener instead?

See also https://stackoverflow.com/a/77166531/281545

__WXOSX__